requirements.txt
pgcli
SQLAlchemy<2.0
psycopg2-binary
pandas
apache-airflow
requests
beautifulsoup4